Reference

PhaseSpaceDTFE.BVH

PhaseSpaceDTFE.BVH

PhaseSpaceDTFE.PS_DTFE

PhaseSpaceDTFE.PS_DTFE_subbox

PhaseSpaceDTFE.SimBox

PhaseSpaceDTFE.PS_DTFE_periodic

PhaseSpaceDTFE.PS_DTFE_periodic

PhaseSpaceDTFE.density

PhaseSpaceDTFE.density_subbox

PhaseSpaceDTFE.findBox

PhaseSpaceDTFE.findCandidateSimplices

PhaseSpaceDTFE.findIntersections

PhaseSpaceDTFE.frame

PhaseSpaceDTFE.frame_velocities

PhaseSpaceDTFE.get_coords_chunk

PhaseSpaceDTFE.get_coords_in_subbox

PhaseSpaceDTFE.get_subbox_estimator

PhaseSpaceDTFE.get_subbox_estimator

PhaseSpaceDTFE.get_subboxes

PhaseSpaceDTFE.inSimplices

PhaseSpaceDTFE.intersection

PhaseSpaceDTFE.numberOfStreams

PhaseSpaceDTFE.numberOfStreams_subbox

PhaseSpaceDTFE.ps_dtfe_subbox

PhaseSpaceDTFE.ps_dtfe_subbox

PhaseSpaceDTFE.translate

PhaseSpaceDTFE.unwrap_s

PhaseSpaceDTFE.unwrap_x_

PhaseSpaceDTFE.velocity

PhaseSpaceDTFE.velocitySum

PhaseSpaceDTFE.velocitySum_subbox

PhaseSpaceDTFE.velocity_subbox

PhaseSpaceDTFE.volume

PhaseSpaceDTFE.BVH — Type

BVH(data, box, depth, left, right)

A Bounding Volume Hierarchy (BVH) tree structure on a set of geometric objects.

PhaseSpaceDTFE.BVH — Method

BVH(data, box::Matrix{Float64}, points::Matrix{Float64}, simplices, depth::Int)

Generates a Bounding Volume Hierarchy (BVH) tree structure.

PhaseSpaceDTFE.PS_DTFE — Type

PS_DTFE(positions_initial, positions, velocities, m, depth, box)

Generates an Phase-Space DTFE object given the initial positions and the final positions of the N-body particles. The Boundary Volume Hirarchy goes depth levels deep. It contains

rho::Vector{Float64}
Drho::Matrix{Float64}
Dv::Array{Float64}
tree::BVH
simplices::Matrix{Int}
positions::Matrix{Float64}
velocities::Matrix{Float64}
positions_initial::Matrix{Float64}

PhaseSpaceDTFE.PS_DTFE_subbox — Type

PS_DTFE_subbox

Phase-Space DTFE subbox object containing

N_sub::Int64
N_target::Int64
m::Float64
depth::Int64
dir::String
L::Float64
Ni::Int64

PhaseSpaceDTFE.SimBox — Type

SimBox(L, Ni)

Sets the size and resolution of the simulation box.

Examples

julia> SimBox(100, 64)
SimBox(100.0, 64)

PhaseSpaceDTFE.PS_DTFE_periodic — Method

PS_DTFE_periodic(coords_q, coords_x, m, depth, sim_box; pad=0.05)

PhaseSpaceDTFE.PS_DTFE_periodic — Method

PS_DTFE_periodic(coords_q, coords_x, velocities, m, depth, sim_box; pad=0.05)

PhaseSpaceDTFE.density — Method

density(p::Vector{Float64}, estimator::PS_DTFE)

Reconstruct the Phase-Space DTFE density in the point p.

PhaseSpaceDTFE.density_subbox — Method

density_subbox(coords_arr, ps_dtfe_sb)

Reconstruct the Phase-Space DTFE density in the point p.

PhaseSpaceDTFE.findBox — Method

findBox(p::Vector{Float64}, BVH_tree::BVH)

Find the box in the Bounding Volume Hierarchy that contains the point p.

PhaseSpaceDTFE.findCandidateSimplices — Method

findCandidateSimplices(p::Vector{Float64}, BVH_tree::BVH)

Given a Bounding Volume Hierarchy, findCandidateSimplices find a set simplices that may contain the point p. The function output indices of candidate simplices.

PhaseSpaceDTFE.findIntersections — Method

findIntersections(p::Vector{Float64}, BVH_tree::BVH, points, simplices)

Find the indices of the simplices that intersect the point p.

PhaseSpaceDTFE.frame — Function

frame(coords_q, coords_x, L, pad=0.05)

Cosmological N-body simulations normally work with periodic boundary conditions. This function added a padding (builds a frame) around the simulation box which helps implement the periodic boundary conditions in the Delaunay tesselation.

PhaseSpaceDTFE.frame_velocities — Function

frame_velocities(coords_x, velocities, L, pad=0.05)

Cosmological N-body simulations normally work with periodic boundary conditions. This function added a padding (builds a frame) of the velocities around the simulation box which helps implement the periodic boundary conditions in the Delaunay tesselation.

PhaseSpaceDTFE.get_coords_chunk — Function

get_coords_chunk(coords, m, m_idx, random=false)

PhaseSpaceDTFE.get_coords_in_subbox — Method

get_coords_in_subbox(coords, idx, N_sub, L)

Obtain the positions in a given subbox.

PhaseSpaceDTFE.get_subbox_estimator — Method

get_subbox_estimator(coords_q, coords_x, velocities, idx, N_sub, m, depth, sim_box::SimBox; pad=0.05, dir="./ps_dtfe")

Obtain the subbox estimator associated to a given position.

PhaseSpaceDTFE.get_subbox_estimator — Method

get_subbox_estimator(coords_q, coords_x, idx, N_sub, m, depth, sim_box::SimBox; pad=0.05, dir="./ps_dtfe")

Obtain the subbox estimator associated to a given position.

PhaseSpaceDTFE.get_subboxes — Method

get_subboxes(ps_dtfe_sub::PS_DTFE_subbox)

PhaseSpaceDTFE.inSimplices — Method

inSimplices(p::Vector{Float64}, estimator::PS_DTFE)

Finds the simplices in the tesselation that include the point p.

PhaseSpaceDTFE.intersection — Method

intersection(p::Vector{Float64}, simplex)

Check whether the point p is included in the simplex.

PhaseSpaceDTFE.numberOfStreams — Method

numberOfStreams(p::Vector{Float64}, estimator::PS_DTFE)

Evaluates the number of streams the point p is in.

PhaseSpaceDTFE.numberOfStreams_subbox — Method

numberOfStreams_subbox(coords_arr, ps_dtfe_sb)

Evaluates the number of streams the point p is in.

PhaseSpaceDTFE.ps_dtfe_subbox — Method

ps_dtfe_subbox(coords_q, coords_x, velocities, m, depth, sim_box::SimBox; N_target=128, pad=0.05, dir="./ps_dtfe")

Generates an Phase-Space DTFE subbox given the initial positions and the final positions and final velocities of the N-body particles. The Boundary Volume Hirarchy goes depth levels deep.

PhaseSpaceDTFE.ps_dtfe_subbox — Method

ps_dtfe_subbox(coords_q, coords_x, m, depth, sim_box::SimBox; N_target=128, pad=0.05, dir="./ps_dtfe")

Generates an Phase-Space DTFE subbox given the initial positions and the final positions of the N-body particles. The Boundary Volume Hirarchy goes depth levels deep.

PhaseSpaceDTFE.translate — Method

translate(coords_q, coords_x, L)

Cosmological N-body simulations normally work with periodic boundary conditions. This function shifts both the intial and final positions of the particles such that they lay in the simulation box in Eulerian space.

PhaseSpaceDTFE.unwrap_s — Method

unwrap_s(s, L)

Cosmological N-body simulations normally work with periodic boundary conditions. This function undoes the wrapping of particles around the box in the displacement field, assuming the particles did not travel more than half the size of the box.

PhaseSpaceDTFE.unwrap_x_ — Method

unwrap_x_(q, x, L)

Cosmological N-body simulations normally work with periodic boundary conditions. This function undoes the wrapping of particles around the box, assuming the particles did not travel more than half the size of the box.

PhaseSpaceDTFE.velocity — Function

velocity(p::Vector{Float64}, estimator::PS_DTFE, single_stream=false)

Reconstructs the Phase-Space DTFE velocities in the point p.

PhaseSpaceDTFE.velocitySum — Method

velocitySum(p::Vector{Float64}, estimator::PS_DTFE)

Reconstructs the mass weighted sum of the Phase-Space DTFE velocities in the point p.

PhaseSpaceDTFE.velocitySum_subbox — Method

velocitySum_subbox(coords_arr, ps_dtfe_sb)

Reconstructs the mass weighted sum of the Phase-Space DTFE velocities in an array of points.

PhaseSpaceDTFE.velocity_subbox — Method

velocity_subbox(coords_arr, ps_dtfe_sb)

Reconstructs the Phase-Space DTFE velocities in an array of points.

PhaseSpaceDTFE.volume — Method

volume(sim, points)

Evaluate the volume of the simplex.